Remote_code/requirements.txt
gradio
transformers==4.39.3
torch
safetensors
xgboost
joblib
pandas
numpy
google-genai
python-dotenv
supabase
scikit-learn
